Moved parsing and searching logic to HexSearchService (HexView.Avalonia)
