+++ HexView/Views/SingleView.axaml.cs
@@ -225,139 +225,6 @@ public partial class SingleView : UserControl
         }
     }
 
-    private static bool TryParseAddress(string? text, out long address)
-    {
-        address = 0;
-        if (string.IsNullOrWhiteSpace(text)) return false;
-        text = text.Trim();
-        if (text.StartsWith("0x", StringComparison.OrdinalIgnoreCase))
-        {
-            return long.TryParse(text.Substring(2), System.Globalization.NumberStyles.HexNumber, null, out address);
-        }
-        // Try hex if contains A-F
-        if (text.IndexOfAny(new[] { 'A','B','C','D','E','F','a','b','c','d','e','f' }) >= 0)
-        {
-            return long.TryParse(text, System.Globalization.NumberStyles.HexNumber, null, out address);
-        }
-        return long.TryParse(text, out address);
-    }
-
-    private static bool TryParseHexBytes(string? text, out byte[] bytes)
-    {
-        bytes = Array.Empty<byte>();
-        if (string.IsNullOrWhiteSpace(text)) return false;
-        text = text.Trim();
-        // Remove spaces and 0x markers
-        var cleaned = new String(text.Where(c => Uri.IsHexDigit(c)).ToArray());
-        if (cleaned.Length < 2 || cleaned.Length % 2 != 0) return false;
-        var result = new byte[cleaned.Length / 2];
-        for (int i = 0; i < result.Length; i++)
-        {
-            if (!byte.TryParse(cleaned.AsSpan(i * 2, 2), System.Globalization.NumberStyles.HexNumber, null, out result[i]))
-            {
-                return false;
-            }
-        }
-        bytes = result;
-        return bytes.Length > 0;
-    }
-
-    private long? FindNextValue(byte[] pattern, long startOffset)
-    {
-        if (HexViewControl1.HexFormatter is null || HexViewControl1.LineReader is null) return null;
-        long length = HexViewControl1.HexFormatter.Length;
-        if (pattern.Length == 0 || length == 0) return null;
-
-        const int chunkSize = 64 * 1024;
-        var buffer = new byte[Math.Max(chunkSize, pattern.Length * 2)];
-        long pos = startOffset;
-        bool wrapped = false;
-
-        while (true)
-        {
-            var read = HexViewControl1.LineReader.Read(pos, buffer, buffer.Length);
-            if (read <= 0)
-            {
-                if (wrapped) break;
-                pos = 0; wrapped = true; continue;
-            }
-            var idx = IndexOf(buffer, read, pattern);
-            if (idx >= 0)
-            {
-                return pos + idx;
-            }
-            pos += Math.Max(1, read - (pattern.Length - 1));
-            if (pos >= length)
-            {
-                if (wrapped) break;
-                pos = 0; wrapped = true;
-            }
-        }
-        return null;
-    }
-
-    private long? FindPrevValue(byte[] pattern, long startOffset)
-    {
-        if (HexViewControl1.HexFormatter is null || HexViewControl1.LineReader is null) return null;
-        long length = HexViewControl1.HexFormatter.Length;
-        if (pattern.Length == 0 || length == 0) return null;
-
-        const int chunkSize = 64 * 1024;
-        var buffer = new byte[Math.Max(chunkSize, pattern.Length * 2)];
-        long pos = Math.Max(0, startOffset - buffer.Length);
-        bool wrapped = false;
-
-        while (true)
-        {
-            var toRead = (int)Math.Min(buffer.Length, startOffset - pos + 1);
-            if (toRead <= 0)
-            {
-                if (wrapped) break;
-                startOffset = length - 1; pos = Math.Max(0, startOffset - buffer.Length); wrapped = true; continue;
-            }
-            var read = HexViewControl1.LineReader.Read(pos, buffer, toRead);
-            if (read <= 0)
-            {
-                if (wrapped) break;
-                startOffset = length - 1; pos = Math.Max(0, startOffset - buffer.Length); wrapped = true; continue;
-            }
-            var idx = LastIndexOf(buffer, read, pattern);
-            if (idx >= 0)
-            {
-                return pos + idx;
-            }
-            if (pos == 0)
-            {
-                if (wrapped) break;
-                startOffset = length - 1; pos = Math.Max(0, startOffset - buffer.Length); wrapped = true; continue;
-            }
-            var step = Math.Max(1, read - (pattern.Length - 1));
-            startOffset = pos - 1; // next window end
-            pos = Math.Max(0, pos - step);
-        }
-        return null;
-    }
-
-    private static int IndexOf(byte[] buffer, int count, byte[] pattern)
-    {
-        if (pattern.Length == 0 || count < pattern.Length) return -1;
-        int lastStart = count - pattern.Length;
-        for (int i = 0; i <= lastStart; i++)
-        {
-            int j = 0; for (; j < pattern.Length; j++) if (buffer[i + j] != pattern[j]) break; if (j == pattern.Length) return i;
-        }
-        return -1;
-    }
-
-    private static int LastIndexOf(byte[] buffer, int count, byte[] pattern)
-    {
-        if (pattern.Length == 0 || count < pattern.Length) return -1;
-        for (int i = count - pattern.Length; i >= 0; i--)
-        {
-            int j = 0; for (; j < pattern.Length; j++) if (buffer[i + j] != pattern[j]) break; if (j == pattern.Length) return i;
-        }
-        return -1;
-    }
 
     private void MoveCaretToOffset(long offset)
     {
@@ -376,7 +243,7 @@ public partial class SingleView : UserControl
 
         if (string.Equals(mode, "Address", StringComparison.OrdinalIgnoreCase))
         {
-            if (TryParseAddress(query, out var addr))
+            if (HexSearchService.TryParseAddress(query, out var addr))
             {
                 var max = Math.Max(0, HexViewControl1.HexFormatter.Length - 1);
                 MoveCaretToOffset(Math.Max(0, Math.Min(max, addr)));
@@ -384,14 +251,15 @@ public partial class SingleView : UserControl
             return;
         }
 
-        if (!TryParseHexBytes(query, out var pattern))
+        if (!HexSearchService.TryParseHexBytes(query, out var pattern))
         {
             return;
         }
 
         var start = HexViewControl1.CaretOffset;
-        long? found = forward ? FindNextValue(pattern, Math.Min(start + 1, HexViewControl1.HexFormatter.Length - 1))
-                              : FindPrevValue(pattern, Math.Max(0, start - 1));
+        long? found = forward
+            ? HexSearchService.FindNextValue(HexViewControl1.LineReader!, HexViewControl1.HexFormatter.Length, pattern, Math.Min(start + 1, HexViewControl1.HexFormatter.Length - 1))
+            : HexSearchService.FindPrevValue(HexViewControl1.LineReader!, HexViewControl1.HexFormatter.Length, pattern, Math.Max(0, start - 1));
         if (found.HasValue)
         {
             MoveCaretToOffset(found.Value);
